refactor: provider 폴더 분리
