+++ src/App.tsx
@@ -5,9 +5,9 @@ import { ComplexForm } from "./components/ComplexForm";
 import { Header } from "./components/Header";
 import { ItemList } from "./components/ItemList";
 import { NotificationSystem } from "./components/NotificationSystem";
-import { NotificationProvider } from "./context/NotificationProvider";
-import { ThemeProvider } from "./context/ThemeProvider";
-import { UserProvider } from './context/UserProvider';
+import { NotificationProvider } from "./provider/NotificationProvider";
+import { ThemeProvider } from "./provider/ThemeProvider";
+import { UserProvider } from "./provider/UserProvider";
 import { generateItems } from "./utils";
 
 const App: React.FC = () => {
@@ -21,26 +21,26 @@ const App: React.FC = () => {
   }, []);
 
   return (
-		<ThemeProvider>
-			<AppContainer>
-				<NotificationProvider>
-					<UserProvider>
-						<Header />
-					</UserProvider>
-					<div className="container mx-auto px-4 py-8">
-						<div className="flex flex-col md:flex-row">
-							<div className="w-full md:w-1/2 md:pr-4">
-								<ItemList items={items} onAddItemsClick={addItems} />
-							</div>
-							<div className="w-full md:w-1/2 md:pl-4">
-								<ComplexForm />
-							</div>
-						</div>
-					</div>
-					<NotificationSystem />
-				</NotificationProvider>
-			</AppContainer>
-		</ThemeProvider>
+    <ThemeProvider>
+      <AppContainer>
+        <NotificationProvider>
+          <UserProvider>
+            <Header />
+          </UserProvider>
+          <div className="container mx-auto px-4 py-8">
+            <div className="flex flex-col md:flex-row">
+              <div className="w-full md:w-1/2 md:pr-4">
+                <ItemList items={items} onAddItemsClick={addItems} />
+              </div>
+              <div className="w-full md:w-1/2 md:pl-4">
+                <ComplexForm />
+              </div>
+            </div>
+          </div>
+          <NotificationSystem />
+        </NotificationProvider>
+      </AppContainer>
+    </ThemeProvider>
   );
 };
