+++ src/provider/ThemeProvider.tsx
@@ -1,21 +1,20 @@
-
 import { PropsWithChildren, useCallback, useState } from "react";
-import { ThemeContext, ThemeContextType } from "./ThemeContext";
+import { ThemeContext, ThemeContextType } from "../context/ThemeContext";
 
 export const ThemeProvider: React.FC<PropsWithChildren> = ({ children }) => {
   const [theme, setTheme] = useState<ThemeContextType["theme"]>("light");
-  
+
   const toggleTheme = useCallback(() => {
     setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
   }, []);
-  
-    const contextValue: ThemeContextType = {
-      theme,
-      toggleTheme,
-    };
+
+  const contextValue: ThemeContextType = {
+    theme,
+    toggleTheme,
+  };
 
   return (
-    <ThemeContext.Provider value= { contextValue }>
+    <ThemeContext.Provider value={ contextValue }>
       { children }
     </ThemeContext.Provider>
   );
