+++ src/provider/UserProvider.tsx
@@ -1,28 +1,25 @@
 import { PropsWithChildren, useState } from "react";
 import { User } from "../types/User";
-import { UserContext } from "./UserContext";
+import { UserContext } from "../context/UserContext";
 
 export const UserProvider: React.FC<PropsWithChildren> = ({ children }) => {
   const [user, setUser] = useState<User | null>(null);
 
-  const login = 
-    (email: string) => {
-      setUser({ id: 1, name: "홍길동", email });
-    }
+  const login = (email: string) => {
+    setUser({ id: 1, name: "홍길동", email });
+  };
 
   const logout = () => {
     setUser(null);
-  }
+  };
 
   const contextValue = {
     user,
     login,
     logout,
-  }
+  };
 
   return (
-    <UserContext.Provider value={contextValue}>
-      {children}
-    </UserContext.Provider>
+    <UserContext.Provider value={contextValue}>{children}</UserContext.Provider>
   );
 };
